requirements.txt
pandas
numpy
scikit-learn
Flask
Flask-cors
kneed
matplotlib
scipy
gunicorn
xgboost
dvc
pyyaml
.